push_swap[ci]: add function ft_populate_stack(), refactore in push_swap.c main() ft_handle_args()
